+++ module-domain/src/main/java/com/kernel360/auth/repository/AuthRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AuthRepository extends JpaRepository <Auth, Id> {
 
-    Auth findOneByMemberNo(Integer memberNo);
+    Auth findOneByMemberNo(Long memberNo);
 
 }
